refactor to put Ownship leg detectors all in one package
