+++ org.mwc.debrief.track_shift/src/org/mwc/debrief/track_shift/zig_detector/IOwnshipLegDetector.java
@@ -2,6 +2,8 @@ package org.mwc.debrief.track_shift.zig_detector;
 
 import java.util.List;
 
+import org.mwc.debrief.track_shift.zig_detector.ownship.LegOfData;
+
 public interface IOwnshipLegDetector
 {
   List<LegOfData> identifyOwnshipLegs(long[] times, double[] rawSpeeds,
