+++ org.mwc.debrief.track_shift/src/org/mwc/debrief/track_shift/zig_detector/ownship/OwnshipLegDetector.java
@@ -1,8 +1,10 @@
-package org.mwc.debrief.track_shift.zig_detector;
+package org.mwc.debrief.track_shift.zig_detector.ownship;
 
 import java.util.ArrayList;
 import java.util.List;
 
+import org.mwc.debrief.track_shift.zig_detector.IOwnshipLegDetector;
+import org.mwc.debrief.track_shift.zig_detector.Precision;
 import org.mwc.debrief.track_shift.zig_detector.moving_average.CenteredMovingAverage;
 import org.mwc.debrief.track_shift.zig_detector.moving_average.TimeBasedMovingAverage;
